修改三大表DF列属性

+++ cash_flow_analysis.py
@@ -5,8 +5,8 @@ def get_sheet_name():
     return '现金流分析'
 
 def profit_quality_data(p_df, c_df):
-    df = c_df[['报表日期', '经营活动产生的现金流量净额']].astype(float)
-    df = df.join(p_df[['五、净利润']].astype(float))
+    df = c_df[['报表日期', '经营活动产生的现金流量净额']]
+    df = df.join(p_df[['五、净利润']])
     # 写入excel
     wbm = wm.Workbook_Manage(get_sheet_name())
     wbm.write_dataframe(df)
@@ -14,8 +14,8 @@ def profit_quality_data(p_df, c_df):
     wbm.write_line_chart('利润质量分析', [1,df.shape[0]+1,2,3],[2,df.shape[0]+1,1],f'F1')
 
 def income_growth_analysis_data(p_df, c_df):
-    df = c_df[['报表日期', '销售商品、提供劳务收到的现金']].astype(float)
-    df = df.join(p_df[['营业收入']].astype(float))
+    df = c_df[['报表日期', '销售商品、提供劳务收到的现金']]
+    df = df.join(p_df[['营业收入']])
     # 写入excel
     wbm = wm.Workbook_Manage(get_sheet_name())
     wbm.write_dataframe(df)
@@ -24,8 +24,8 @@ def income_growth_analysis_data(p_df, c_df):
     wbm.write_line_chart('营业收入增长分析', [start,df.shape[0]+start,2,3],[2,df.shape[0]+1,1],f'F{start}')
 
 def cash_support_analysis_data(c_df, b_df):
-    df = b_df[['报表日期', '短期借款', '交易性金融负债', '一年内到期的非流动负债', '长期借款', '应付债券']].astype(float)
-    df = df.join(c_df[['现金的期末余额', '投资活动现金流出小计', '分配股利、利润或偿付利息所支付的现金']].astype(float))
+    df = b_df[['报表日期', '短期借款', '交易性金融负债', '一年内到期的非流动负债', '长期借款', '应付债券']]
+    df = df.join(c_df[['现金的期末余额', '投资活动现金流出小计', '分配股利、利润或偿付利息所支付的现金']])
     df['有息负债'] = df['短期借款'] + df['交易性金融负债'] + df['一年内到期的非流动负债'] + df['长期借款'] + df['应付债券']
     # 写入excel
     wbm = wm.Workbook_Manage(get_sheet_name())
